ext4: improve error recovery code paths in __ext4_remount()
commit 4c0b4818b1f636bc96359f7817a2d8bab6370162 upstream.

If there are failures while changing the mount options in
__ext4_remount(), we need to restore the old mount options.

This commit fixes two problem.  The first is there is a chance that we
will free the old quota file names before a potential failure leading
to a use-after-free.  The second problem addressed in this commit is
if there is a failed read/write to read-only transition, if the quota
has already been suspended, we need to renable quota handling.

Cc: stable@kernel.org
Link: https://lore.kernel.org/r/20230506142419.984260-2-tytso@mit.edu
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Change-Id: Ia7368fafcced60be10f21c8ba4dd044fb8971157
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
[bananafunction: minor adaption: MS_RDONLY instead of SB_RDONLY]
Signed-off-by: bananafunction <terathree2@web.de>
(cherry picked from commit 013b491f63eca342283da4297aded939533e62d1)
Signed-off-by: TogoFire <togofire@mailfence.com>